requirements.txt
py-cord
python-dotenv
PyNaCl
gTTS
faster-whisper
nvidia-cublas-cu11
nvidia-cudnn-cu11
gpt4all